requirements.txt
Django==4.0.5
django-extensions==3.1.5
djangorestframework==3.13.1
dr-scaffold==2.1.2
python-decouple==3.6
